refactored reminderService
