+++ src/main/resources/newCallNotificationEmail.mjml
@@ -1,0 +1,151 @@
+<mjml>
+  <mj-head>
+    <mj-style>
+      .box-shadow td {
+        box-shadow: 0 4px 6px hsla(0, 0%, 46%, 0.2), 0 1px 3px hsla(0, 0%, 46%, 0.2);
+      }
+    </mj-style>
+  </mj-head>
+  <mj-body>
+    <mj-section>
+      <mj-column width="100px" vertical-align="middle">
+        <mj-image src="https://cclcalls.org/img/ccl-logo-thumbnail.png"></mj-image>
+      </mj-column>
+      <mj-column width="100%" vertical-align="middle">
+        <mj-text font-size="24px">CCL's Monthly Calling Campaign</mj-text>
+      </mj-column>
+      <mj-column width="100%">
+        <mj-divider border-color="#000"></mj-divider>
+      </mj-column>
+    </mj-section>
+    <mj-hero background-color="white" padding="0px">
+      <mj-text font-size="14px" line-height="14px">
+        Dear {CallerName}
+      </mj-text>
+      <mj-text padding="20px" font-size="14px" line-height="14px">
+        It's time to call your Member of Congress about climate change. Please find 5 minutes in the next day or
+        two to urge climate action.
+      </mj-text>
+      <mj-divider border-color="#000"></mj-divider>
+    </mj-hero>
+    <mj-hero background-color="white" padding="0px">
+      <mj-text padding="10px" align="center" font-size="20px" line-height="20px">
+        Call #1 - {MOC1Name} - {MOC1District}
+      </mj-text>
+      <mj-text align="center" font-size="14px" line-height="14px">
+        {MOC1Number}
+      </mj-text>
+      <mj-text font-size="14px" padding-bottom="5px" line-height="14px">
+        Recommended Script:
+      </mj-text>
+      <mj-text font-size="14px" padding-top="5px" padding-bottom="5px" padding-left="40px" line-height="14px">
+        Intro: Hello, my name is {CallerName}, and I live in {MOC1District}. I am calling about climate change.
+      </mj-text>
+      <mj-text font-size="14px" padding-top="5px" padding-left="40px" line-height="14px">
+        Request: {MOC1RequestScript}
+      </mj-text>
+      <mj-button font-size="24px" color="white" background-color="#4c4cff" href="{MOC1ICalled}" css-class="box-shadow">
+        I called {MOC1Name}!
+      </mj-button>
+      <mj-text font-size="14px" line-height="14px">
+        For more information and talking points relating to {MOC1Name} view the <a href='{MOC1Guide}'> Call-In
+          Guide</a>
+      </mj-text>
+      <mj-divider border-color="#000"></mj-divider>
+    </mj-hero>
+    <mj-hero background-color="white" padding="0px">
+      <mj-text padding="10px" align="center" font-size="20px" line-height="20px">
+        Call #2 - {MOC2Name} - {MOC2District}
+      </mj-text>
+      <mj-text align="center" font-size="14px" line-height="14px">
+        {MOC2Number}
+      </mj-text>
+      <mj-text font-size="14px" padding-bottom="5px" line-height="14px">
+        Recommended Script:
+      </mj-text>
+      <mj-text font-size="14px" padding-top="5px" padding-bottom="5px" padding-left="40px" line-height="14px">
+        Intro: Hello, my name is {CallerName}, and I live in {MOC2District}. I am calling about climate change.
+      </mj-text>
+      <mj-text font-size="14px" padding-top="5px" padding-left="40px" line-height="14px">
+        Request: {MOC2RequestScript}
+      </mj-text>
+      <mj-button font-size="24px" color="white" background-color="#4c4cff" href="{MOC2ICalled}" css-class="box-shadow">
+        I called {MOC2Name}!
+      </mj-button>
+      <mj-text font-size="14px" line-height="14px">
+        For more information and talking points relating to {MOC1Name} view the <a href='{MOC2Guide}'> Call-In
+          Guide</a>
+      </mj-text>
+      <mj-divider border-color="#000"></mj-divider>
+    </mj-hero>
+    <mj-hero background-color="white" padding="0px">
+      <mj-text padding="10px" align="center" font-size="20px" line-height="20px">
+        Call #3 - {MOC3Name} - {MOC3District}
+      </mj-text>
+      <mj-text align="center" font-size="14px" line-height="14px">
+        {MOC3Number}
+      </mj-text>
+      <mj-text font-size="14px" padding-bottom="5px" line-height="14px">
+        Recommended Script:
+      </mj-text>
+      <mj-text font-size="14px" padding-top="5px" padding-bottom="5px" padding-left="40px" line-height="14px">
+        Intro: Hello, my name is {CallerName}, and I live in {MOC3District}. I am calling about climate change.
+      </mj-text>
+      <mj-text font-size="14px" padding-top="5px" padding-left="40px" line-height="14px">
+        Request: {MOC3RequestScript}
+      </mj-text>
+      <mj-button font-size="24px" color="white" background-color="#4c4cff" href="{MOC3ICalled}" css-class="box-shadow">
+        I called {MOC3Name}!
+      </mj-button>
+      <mj-text font-size="14px" line-height="14px">
+        For more information and talking points relating to {MOC1Name} view the <a href='{MOC3Guide}'> Call-In
+          Guide</a>
+      </mj-text>
+      <mj-divider border-color="#000"></mj-divider>
+    </mj-hero>
+    <mj-hero background-color="white" padding="0px">
+      <mj-text padding="10px" align="center" font-size="20px" line-height="20px">
+        Once you have called each of your Members of Congress, invite your friends to do the same.
+      </mj-text>
+      <mj-button font-size="24px" color="white" background-color="#4c4cff" href="{IInvited}" css-class="box-shadow">
+        I invited a friend to join the Monthly Call-In Campaign!
+      </mj-button>
+      <mj-divider border-color="#000"></mj-divider>
+    </mj-hero>
+    <mj-section>
+      <mj-column>
+        <mj-text font-size="14px">
+          Comments, questions, moved districts?
+        </mj-text>
+        <mj-text font-size="14px">
+          Reply to this email with your feedback, questions, and/or district change request.
+        </mj-text>
+        <mj-text font-size="14px">
+          What is the Monthly Calling Campaign?
+        </mj-text>
+        <mj-text font-size="14px">
+          The Monthly Calling Campaign is an effort by Citizens' Climate Lobby to inspire as many people as
+          possible to call their Member of Congress and express their concern about climate change.
+          <br /><br /> Congressional staffers and representatives have told us nothing is more powerful in
+          helping a representative know what their constituents want than written letters and phone calls.
+        </mj-text>
+        <mj-text font-size="14px">
+          What is CCL?
+        </mj-text>
+        <mj-text font-size="14px">
+          CCL exists to create the political will for climate solutions by enabling individual breakthroughs
+          in the exercise of personal and political power.
+          <br /><br />You can learn more at <a href="https://citizensclimatelobby.org/about-ccl/">citizensclimatelobby.org</a>
+        </mj-text>
+      </mj-column>
+    </mj-section>
+    <mj-section>
+      <mj-column>
+        <mj-text align="center">
+          From your friends at <a href="citizensclimatelobby.org">CCL</a>. To unsubscribe, simply reply to
+          this email with "unsubscribe" in the body.
+        </mj-text>
+      </mj-column>
+    </mj-section>
+  </mj-body>
+</mjml>
